Turn all doubles to floats
[ci skip]

+++ yampa/src/FRP/Yampa/InternalCore.hs
@@ -105,12 +105,12 @@ import FRP.Yampa.Event       (Event (..))
 -- agreed reference point in time.
 
 --  Conceptually, Time = R, i.e. time can be 0 -- or even negative.
-type Time = Double      -- [s]
+type Time = Float -- [s]
 
 -- | DTime is the time type for lengths of sample intervals. Conceptually,
 -- DTime = R+ = { x in R | x > 0 }. Don't assume Time and DTime have the
 -- same representation.
-type DTime = Double     -- [s]
+type DTime = Float -- [s]
 
 -- | Signal function that transforms a signal carrying values of some type 'a'
 -- into a signal carrying values of some type 'b'. You can think of it as
